Move object size check to storage protocol
